crm_backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
psycopg2-binary==2.9.10
python-dotenv==1.0.1
bcrypt==4.2.0
python-jose[cryptography]==3.3.0
cryptography>=42.0.0
httpx>=0.27.0
pydantic==2.10.2
pydantic-settings==2.6.1
cloudinary==1.41.0
python-multipart==0.0.12
